Add Sd Clang settings to vendor. This is pulled from TeamExodus device qcom common. the origianl commits it was pulled from
common: add support for SDCLANG
This commit adds the needed bits for SDCLANG support.

SDCLANG is initialised as false currently,
however the implementation allows to override it from device.

The Path is currenly hardcoded to "prebuilts/clang/host/linux-x86/sdclang-3.8/bin"
since we cannot provide support for darwin currently.

The LTO definitions are added and the variable which
gets picked up by build system automatically is set.

Also a new configuration is introduced: SDCLANG_COMMON_FLAGS
These flags will be picked up for arm and arm64 targets.
We are using -O3 currently since SDCLANG defines the levels
of optimisation differently and O3 is a very safe level to use here.

The new configuration SDCLANG_FORCED is initialised
at false and can be overriden. It can be set
to true in order to force SDCLANG to be used
under the condition that it is requested.
Currently we don't want to do this globally,
however it might be useful for OEM builds.

Change-Id: I4f42be82a8a113fdf6406d9879d5586968030517
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

common: sdclang: LTO: filter new no-erratum-fix flag
* We are using this one in build now too
* But it isn't supported by QCLD either
-> filter it out

Change-Id: I148d82f94dad67fdca286d03b8b71fd3846c87a3
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

common: sdclang: Enable by default!
The Dragon unleashes the fire.

Change-Id: Ie00fffc245f83731324ab8a8de3469e54d8da3a4
Signed-off-by: Alex Naidis <alex.naidis@linux.com>